+++ changelog.md
@@ -5,5 +5,6 @@
 - Added more Russian translations.
 - Fixed profit calculations.
 - Switched GitHub Pages to Firebase.
+- Reduced power of "Samson 45K-FGXL".
 
 ## v0.9 - MVP Pre Release - 30/01/2022